requirements.txt
tensorflow==2.15.0
numpy==1.26.4
scikit-learn==1.4.0
opencv-python==4.9.0.80
flask==3.0.2
pillow==10.2.0
matplotlib==3.8.2
gunicorn==21.2.0
flask-cors==4.0.0
